requirements.txt
streamlit==1.3.1
streamlit-chat
Pillow==9.4.0
matplotlib==3.6.2
